Removed clippy warnings

+++ src/lib.rs
@@ -1,13 +1,10 @@
+use std::collections::VecDeque;
 /// A mock for serial communication using the embedded-io-async traits Read and Write.
 ///
 // TODO Change the name of the crate to embedded-io-async-mock
 // TODO Add some module documentation here including examples
 // TODO Check the formatting of the error meesages againt embedded-hal-mock
 use std::vec::Vec;
-use std::{collections::VecDeque, os::unix::thread};
-
-use embedded_io::{Error, ErrorKind, ErrorType};
-use embedded_io_async::{Read, Write};
 
 pub struct Mock {
     // transactions: Vec<SerialTransaction>,
@@ -42,7 +39,7 @@ impl embedded_io_async::Read for Mock {
         match self.transactions.pop_front() {
             Some(Transaction::Read(data)) => {
                 buf.copy_from_slice(&data);
-                Ok(data.len() as usize)
+                Ok(data.len())
             }
             Some(other_transaction) => {
                 self.transactions_aborted = true;
